Added user authentication using a custom user class and manager.
Added some python scripts in the scripts folder
which are used to load data into the Car model.
Used the django-extensions package to run those scripts.
'python manage.py runscript <file-name>'

Added a forms.py file in which there is a custom
user creation form and a custom user change form.

Made changes to the way car images are stored.
Images of cars are now stored in the database
as an ImageField datatype. Any uploaded image will be stored
in the media folder in the root project directory along the
manage.py file.